fix: Check content and associated data exists before scraping

+++ apis/aylo/aylo-scraper.ps1
@@ -151,13 +151,17 @@ function Get-AyloActorJson {
 # Get data for a piece of content
 function Get-AyloJson {
     param (
-        [Parameter(Mandatory)][ValidateSet('actor', 'gallery', 'movie', 'scene', 'serie', 'trailer')][String]$apiType,
+        [Parameter(Mandatory)][ValidateSet('gallery', 'movie', 'scene', 'serie', 'trailer')][String]$apiType,
         [Parameter(Mandatory)][Int]$contentID,
         [Parameter(Mandatory)][String]$pathToUserConfig
     )
     Write-Host `n"Starting scrape for $apiType #$contentID." -ForegroundColor Cyan
 
     $userConfig = Get-Content $pathToUserConfig -raw | ConvertFrom-Json
+    $contentDir = $userConfig.general.contentDirectory
+    $contentDownloadDir = $userConfig.general.contentDownloadDirectory
+    $dataDir = $userConfig.general.dataDirectory
+    $dataDownloadDir = $userConfig.general.dataDownloadDirectory
 
     # Attempt to scrape content data
     $result = Get-AyloQueryData -apiType $apiType -contentID $contentID -pathToUserConfig $pathToUserConfig
@@ -169,48 +173,56 @@ function Get-AyloJson {
     $parentStudio = $result.brandMeta.displayName
     if ($result.collections.count -gt 0) { $studio = $result.collections[0].name }
     else { $studio = $parentStudio }
-
-    # Skip creating JSON if the content already exists in either the download or
-    # storage directory
-    $topDirs = @($userConfig.general.contentDownloadDirectory, $userConfig.general.contentDirectory)
     $subDir = Join-Path "aylo" $apiType $parentStudio $studio
 
-    foreach ($topDir in $topDirs) {
-        $contentDir = Join-Path $topDir $subDir
-        
-        # Skip creating JSON if the content already exists
-        if (Test-Path -LiteralPath $contentDir) {
-            $contentFile = Get-ChildItem $contentDir | Where-Object { $_.BaseName -match "^$contentID\s" }
-            if ($contentFile.Length -gt 0) {
-                Write-Host "Media already exists at $($contentFile.FullName). Skipping JSON generation for $apiType #$contentID."
-
-                # Return the path to the existing JSON file
-                $title = Get-SanitizedTitle -title $result.title
-                $filename = "$contentID $title.json"
-                $pathToExistingJson = Join-Path $userConfig.general.dataDownloadDirectory $subDir $filename
-                return $pathToExistingJson
+    # Skip creating JSON if both the JSON and the content already exist in either directory
+    $existingPath = $null
+    $existingJson = $null
+    foreach ($dir in @($contentDir, $contentDownloadDir)) {
+        $testPath = Join-Path $dir $subDir
+        if (Test-Path -LiteralPath $testPath) {
+            $filename = Get-ChildItem -LiteralPath $testPath | Where-Object { $_.BaseName -match "^$contentID\s" }
+            if ($null -ne $filename -and (Test-Path -LiteralPath $filename.FullName)) {
+                # Check the associated JSON also exists
+                foreach ($dDir in @($dataDir, $dataDownloadDir)) {
+                    $dataTestPath = Join-Path $dDir $subDir
+                    if (Test-Path -LiteralPath $dataTestPath) {
+                        $jsonFilename = Get-ChildItem -LiteralPath $dataTestPath | Where-Object { $_.BaseName -match "^$contentID\s" }
+                        if ($null -ne $jsonFilename -and (Test-Path -LiteralPath $jsonFilename.FullName)) {
+                            # Check the file exists in the directory
+                            $existingPath = $filename.FullName
+                            $existingJson = $jsonFilename.FullName
+                        }
+                    }
+                }
             }
         }
     }
 
-    # Output the JSON file
-    $title = Get-SanitizedTitle -title $result.title
-    $filename = "$contentID $title.json"
-    $outputDir = Join-Path $userConfig.general.dataDownloadDirectory $subDir
-    if (!(Test-Path -LiteralPath $outputDir)) { New-Item -ItemType "directory" -Path $outputDir }
-    $outputDest = Join-Path $outputDir $filename
+    if ($null -eq $existingPath -or $null -eq $existingJson) {
+        # Output the JSON file
+        $title = Get-SanitizedTitle -title $result.title
+        $filename = "$contentID $title.json"
+        $outputDir = Join-Path $userConfig.general.dataDownloadDirectory $subDir
+        if (!(Test-Path -LiteralPath $outputDir)) { New-Item -ItemType "directory" -Path $outputDir }
+        $outputDest = Join-Path $outputDir $filename
 
-    Write-Host "Generating JSON: $filename"
-    $result | ConvertTo-Json -Depth 32 | Out-File -FilePath $outputDest
+        Write-Host "Generating JSON: $filename"
+        $result | ConvertTo-Json -Depth 32 | Out-File -FilePath $outputDest
 
-    if (!(Test-Path -LiteralPath $outputDest)) {
-        Write-Host "ERROR: $apiType JSON generation failed - $outputDest" -ForegroundColor Red
-        return $null
-    }  
+        if (!(Test-Path -LiteralPath $outputDest)) {
+            Write-Host "ERROR: $apiType JSON generation failed - $outputDest" -ForegroundColor Red
+            return $null
+        }  
+        else {
+            Write-Host "SUCCESS: $apiType JSON generated - $outputDest" -ForegroundColor Green
+            return $outputDest
+        }  
+    }
     else {
-        Write-Host "SUCCESS: $apiType JSON generated - $outputDest" -ForegroundColor Green
-        return $outputDest
-    }  
+        Write-Host "Media already exists at $($existingPath). Skipping JSON generation for $apiType #$contentID."
+        return $existingJson
+    }
 }
 
 # Get data for content related to the given Aylo gallery and output it to a JSON
